Support MaxLinks and MaxHAMTFanout
This introduces support for setting DAG-shaping options that control the maximum number
of links in that a DAG node can have.

It is supported in two ways: first, via `Import` configuration options; second, via `--max-links` and `--max-hamt-fanout` options `ipfs add`.

The resulting DAGs will respect MaxLinks (for files and basic directories) and
MaxHAMTFanout for HAMT directories. When options are unset, the previous Kubo
defaults will apply.