Provide according to strategy
This updates boxo to this PR https://github.com/ipfs/boxo/pull/976, which decentralizes the providing responsibilities.

Then it perfoms the necessary changes in Kubo, which consist in initializing the Pinner, MFS and the blockstore with the provider.System.

Since the provider.System is created first, the reproviding KeyChanFunc is set
later when we can create it.