Provider changes: ensure added blocks get provided according to strategy
This commit ensure that added blocks get provided with "pinned" strategy.

Normally, blocks would get provided at the blockstore or pinner facility, but
when adding with "pinned" strategy, the blockstore does not provide, and the
pinner does not traverse the DAG so we need to provide directly from the Adder.

This is resolved by wrapping the DAGService in a "providingDAGService" which
provides every added block, when using the "pinned" strategy. There are
alternatives like wiring a provider in coreunix but this seems simple enough
and follows existing style as the DAGService is already wrapped as a Syncer.

CoreAPI also overwrote the provider into a NoopProvider when using
--offline. However, we have strategies to follow and it did not overwrite the
blockstore. So --offline would provide anyways with strategy "all", and not
provide with strategy "pinned".

It seems reasonable to provide following the provider strategy when the node
is online, regardless of the client flag. Otherwise we can change it later by
overwriting the provider and the blockstore.

cli/provider tests have been fixed and made to not rely on `--offline`.